issueの修正２

+++ app/controllers/prototypes_controller.rb
@@ -25,7 +25,7 @@ class PrototypesController < ApplicationController
   end
 
   def edit
-    unless user_signed_in?
+    unless user_signed_in? && current_user == @prototype.user
       redirect_to action: :index
     end
     @prototype = Prototype.find(params[:id])
@@ -42,6 +42,9 @@ class PrototypesController < ApplicationController
   end
 
   def destroy
+    unless user_signed_in? && current_user == @prototype.user
+      redirect_to action: :index
+    end
     prototype = Prototype.find(params[:id])
     prototype.destroy
     redirect_to root_path
